Migrate express-zod client to unified typescript generator; move spec to typescript-express-zod/petstore.yaml
Excludes regenerated samples/ output (committed separately).
