+++ bin/configs/typescript-express-zod-client.yaml
@@ -1,5 +1,6 @@
-generatorName: typescript-express-zod-client
+generatorName: typescript
 outputDir: samples/client/petstore/typescript-express-zod-client/builds/default
-inputSpec: modules/openapi-generator/src/test/resources/3_0/typescript-express-zod.yaml
+inputSpec: modules/openapi-generator/src/test/resources/3_0/typescript-express-zod/petstore.yaml
 additionalProperties:
+  framework: express-zod
   hideGenerationTimestamp: "true"
