+++ bin/configs/typescript-express-zod-server.yaml
@@ -1,5 +1,5 @@
 generatorName: typescript-express-zod-server
 outputDir: samples/server/petstore/typescript-express-zod-server/builds/default
-inputSpec: modules/openapi-generator/src/test/resources/3_0/typescript-express-zod.yaml
+inputSpec: modules/openapi-generator/src/test/resources/3_0/typescript-express-zod/petstore.yaml
 additionalProperties:
   hideGenerationTimestamp: "true"
