+++ modules/openapi-generator/src/main/resources/META-INF/services/org.openapitools.codegen.CodegenConfig
@@ -155,6 +155,7 @@ org.openapitools.codegen.languages.TypeScriptClientCodegen
 org.openapitools.codegen.languages.TypeScriptAngularClientCodegen
 org.openapitools.codegen.languages.TypeScriptAureliaClientCodegen
 org.openapitools.codegen.languages.TypeScriptAxiosClientCodegen
+org.openapitools.codegen.languages.TypeScriptExpressZodServerCodegen
 org.openapitools.codegen.languages.TypeScriptFetchClientCodegen
 org.openapitools.codegen.languages.TypeScriptInversifyClientCodegen
 org.openapitools.codegen.languages.TypeScriptJqueryClientCodegen
@@ -166,6 +167,3 @@ org.openapitools.codegen.languages.TypeScriptRxjsClientCodegen
 org.openapitools.codegen.languages.WsdlSchemaCodegen
 org.openapitools.codegen.languages.XojoClientCodegen
 org.openapitools.codegen.languages.ZapierClientCodegen
-
-org.openapitools.codegen.languages.TypescriptExpressZodServerCodegen
-org.openapitools.codegen.languages.TypescriptExpressZodClientCodegen
