--- modules/openapi-generator/src/main/resources/typescript-express-zod-client/api.mustache
@@ -1,1 +1,0 @@
-
