--- modules/openapi-generator/src/main/resources/typescript-express-zod-client/model.mustache
@@ -1,1 +1,0 @@
-
