+++ modules/openapi-generator/src/main/resources/typescript/express-zod/types.mustache
@@ -1,5 +1,5 @@
 /**
- * This code was generated by openapi-generator (typescript-express-zod-client).
+ * This code was generated by openapi-generator (typescript, framework 'express-zod').
  *
  * Changes to this file may cause incorrect behavior and will be lost if
  * the code is regenerated.
