+++ modules/openapi-generator/src/test/java/org/openapitools/codegen/typescript/express/zod/TypeScriptExpressZodClientFrameworkTest.java
@@ -1,0 +1,95 @@
+package org.openapitools.codegen.typescript.express.zod;
+
+import org.openapitools.codegen.ClientOptInput;
+import org.openapitools.codegen.DefaultGenerator;
+import org.openapitools.codegen.SupportingFile;
+import org.openapitools.codegen.TestUtils;
+import org.openapitools.codegen.config.CodegenConfigurator;
+import org.openapitools.codegen.languages.TypeScriptClientCodegen;
+import org.testng.Assert;
+import org.testng.annotations.Test;
+
+import java.io.File;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
+
+/**
+ * Tests for the {@code express-zod} framework of the unified {@code typescript} generator
+ * (the client counterpart of {@code typescript-express-zod-server}).
+ */
+public class TypeScriptExpressZodClientFrameworkTest {
+
+    @Test(description = "express-zod framework replaces the default template set with its supporting files")
+    public void processOptsRegistersExpressZodFiles() {
+        final TypeScriptClientCodegen codegen = new TypeScriptClientCodegen();
+        codegen.additionalProperties().put("framework", "express-zod");
+        codegen.processOpts();
+
+        Set<String> outputs = codegen.supportingFiles().stream()
+                .map(sf -> sf.getFolder().isEmpty()
+                        ? sf.getDestinationFilename()
+                        : sf.getFolder() + "/" + sf.getDestinationFilename())
+                .collect(Collectors.toSet());
+        Assert.assertEquals(outputs, Set.of(
+                "types.ts", "schemas.ts", "http-client.ts", "index.ts",
+                "dtos/types.ts", "dtos/mappers.ts", "dtos/index.ts"),
+                "unexpected supporting files: " + outputs);
+        Assert.assertTrue(codegen.modelTemplateFiles().isEmpty());
+        Assert.assertTrue(codegen.apiTemplateFiles().isEmpty());
+        Assert.assertTrue(codegen.apiDocTemplateFiles().isEmpty());
+    }
+
+    @Test(description = "express-zod framework generates the express-zod client layout end to end")
+    public void endToEndGeneration() throws Exception {
+        final File output = Files.createTempDirectory("typescriptexpresszodclient_").toFile();
+        output.deleteOnExit();
+
+        final CodegenConfigurator configurator = new CodegenConfigurator()
+                .setGeneratorName("typescript")
+                .addAdditionalProperty("framework", "express-zod")
+                .setInputSpec("src/test/resources/3_0/typescript-express-zod/petstore.yaml")
+                .setOutputDir(output.getAbsolutePath().replace("\\", "/"));
+
+        final ClientOptInput clientOptInput = configurator.toClientOptInput();
+        final DefaultGenerator generator = new DefaultGenerator();
+        final List<File> files = generator.opts(clientOptInput).generate();
+        files.forEach(File::deleteOnExit);
+
+        for (String expected : new String[]{
+                "types.ts", "schemas.ts", "http-client.ts", "index.ts",
+                "dtos/types.ts", "dtos/mappers.ts", "dtos/index.ts"}) {
+            Assert.assertTrue(new File(output, expected).exists(), "missing " + expected);
+        }
+        for (String unexpected : new String[]{"apis", "models", "package.json", "http"}) {
+            Assert.assertFalse(new File(output, unexpected).exists(), "unexpected " + unexpected);
+        }
+
+        // Response schemas resolved via TypeScriptExpressZodSupport (guard regression)
+        TestUtils.assertFileContains(Paths.get(output + "/schemas.ts"), "ResponseSchema");
+        TestUtils.assertFileContains(Paths.get(output + "/http-client.ts"),
+                "generated by openapi-generator (typescript, framework 'express-zod')");
+    }
+
+    @Test(description = "the default fetch-api framework is unaffected by the express-zod hooks")
+    public void defaultFrameworkStillGeneratesFetchApi() throws Exception {
+        final File output = Files.createTempDirectory("typescriptdefaultframework_").toFile();
+        output.deleteOnExit();
+
+        final CodegenConfigurator configurator = new CodegenConfigurator()
+                .setGeneratorName("typescript")
+                .setInputSpec("src/test/resources/3_0/typescript-express-zod/petstore.yaml")
+                .setOutputDir(output.getAbsolutePath().replace("\\", "/"));
+
+        final ClientOptInput clientOptInput = configurator.toClientOptInput();
+        final DefaultGenerator generator = new DefaultGenerator();
+        final List<File> files = generator.opts(clientOptInput).generate();
+        files.forEach(File::deleteOnExit);
+
+        Assert.assertTrue(new File(output, "http/isomorphic-fetch.ts").exists(),
+                "default framework should still emit the fetch-api http lib");
+        Assert.assertTrue(new File(output, "package.json").exists());
+    }
+}
