--- modules/openapi-generator/src/test/java/org/openapitools/codegen/typescript/express/zod/TypescriptExpressZodServerCodegenTest.java
@@ -1,19 +1,0 @@
-package org.openapitools.codegen.typescript.express.zod;
-
-import org.openapitools.codegen.*;
-import org.openapitools.codegen.languages.TypescriptExpressZodServerCodegen;
-import io.swagger.models.*;
-import io.swagger.parser.SwaggerParser;
-import org.testng.Assert;
-import org.testng.annotations.Test;
-
-public class TypescriptExpressZodServerCodegenTest {
-
-    TypescriptExpressZodServerCodegen codegen = new TypescriptExpressZodServerCodegen();
-
-    @Test
-    public void shouldSucceed() throws Exception {
-        // TODO: Complete this test.
-        Assert.fail("Not implemented.");
-    }
-}
