+++ modules/openapi-generator/src/test/resources/3_0/typescript-express-zod/petstore.yaml
